Move some classes up the dependency tree to reduce the amount of dependencies for ihmc-perception, which is meant to be lower level.
