+++ atlas/src/main/java/us/ihmc/atlas/multisenseTestbench/MultisenseTestBenchWithZeroPoseModuleNetworkProcessor.java
@@ -23,7 +23,7 @@ import us.ihmc.communication.util.NetworkPorts;
 import us.ihmc.euclid.transform.RigidBodyTransform;
 import us.ihmc.humanoidRobotics.kryo.IHMCCommunicationKryoNetClassList;
 import us.ihmc.humanoidRobotics.kryo.PPSTimestampOffsetProvider;
-import us.ihmc.perception.time.AlwaysZeroOffsetPPSTimestampOffsetProvider;
+import us.ihmc.avatar.networkProcessor.time.AlwaysZeroOffsetPPSTimestampOffsetProvider;
 import us.ihmc.mecano.multiBodySystem.interfaces.OneDoFJointBasics;
 import us.ihmc.pubsub.DomainFactory.PubSubImplementation;
 import us.ihmc.robotModels.FullHumanoidRobotModel;
