+++ ihmc-avatar-interfaces/src/main/java/us/ihmc/avatar/networkProcessor/time/DRCROSAlwaysZeroOffsetPPSTimestampOffsetProvider.java
@@ -2,7 +2,6 @@ package us.ihmc.avatar.networkProcessor.time;
 
 
 import us.ihmc.avatar.ros.DRCROSPPSTimestampOffsetProvider;
-import us.ihmc.perception.time.AlwaysZeroOffsetPPSTimestampOffsetProvider;
 import us.ihmc.utilities.ros.RosNodeInterface;
 
 public class DRCROSAlwaysZeroOffsetPPSTimestampOffsetProvider extends AlwaysZeroOffsetPPSTimestampOffsetProvider implements DRCROSPPSTimestampOffsetProvider
